+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add `mcr generate` command to print a new random master key.
+
 ## [0.1.0] - 2026-06-24
 
 - Initial release